// src/types/auth.ts
export type User = {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  createdAt: string;
  is_verified: boolean;
  role: 'admin' | 'user' | 'super-admin';
  phone: string;
  dial_code: string;
  profile_picture: string;
  authority: string[];
  bio: string | null;
  is_banned: boolean;
  is_deleted: boolean;
  stripe_connect_id: string | null;
  subscription_status: 'active' | 'inactive' | 'expired' | null;
  auth_provider: 'email';
};